Update getKey code and test
