+++ wrapped-keys/nodejs/package.json
@@ -8,8 +8,8 @@
     "test:import": "npx @dotenvx/dotenvx run -- mocha test/**/importKey.spec.ts",
     "test:export": "npx @dotenvx/dotenvx run -- mocha test/**/exportWrappedKey.spec.ts",
     "test:generate": "npx @dotenvx/dotenvx run -- mocha test/**/generateWrappedKey.spec.ts",
-    "test:getMetadata": "npx @dotenvx/dotenvx run -- mocha test/**/getWrappedKeyMetadata.spec.ts",
-    "test:storeMetadata": "npx @dotenvx/dotenvx run -- mocha test/**/storeWrappedKeyMetadata.spec.ts",
+    "test:getKey": "npx @dotenvx/dotenvx run -- mocha test/**/getWrappedKey.spec.ts",
+    "test:storeKey": "npx @dotenvx/dotenvx run -- mocha test/**/storeWrappedKey.spec.ts",
     "test:sign:tx": "npx @dotenvx/dotenvx run -- mocha test/**/signTransactionWithWrappedKey.spec.ts",
     "test:sign:msg": "npx @dotenvx/dotenvx run -- mocha test/**/signMessageWithWrappedKey.spec.ts",
     "test:signAndSend:tx": "npx @dotenvx/dotenvx run -- mocha test/**/signAndSendTransactionWithWrappedKey.spec.ts"
